requirements.txt
torch==1.5.1
torchvision==0.6.1
numpy==1.17
pandas
sklearn
networkx
seaborn
scipy==1.4
scanpy==1.5.1
umap-learn==0.4.6
scvi-tools==0.9.0
intercode
